image_rec/requirements.txt
Pillow~=7.2.0
Flask~=1.1.2
tensorflow>=2.0.0
numpy~=1.19.2